+++ packages/jaspr/lib/src/server/adapters/client_component_adapter.dart
@@ -1,81 +1,100 @@
 import '../../dom/validator.dart';
-import '../../foundation/constants.dart';
 import '../../framework/framework.dart';
+import '../components/client_component_registry.dart';
 import '../markup_render_object.dart';
 import '../options.dart';
-import '../server_binding.dart';
-import 'client_script_adapter.dart';
 import 'element_boundary_adapter.dart';
+import 'server_component_adapter.dart';
 
 class ClientComponentAdapter extends ElementBoundaryAdapter {
-  ClientComponentAdapter(this.name, this.data, super.element);
+  ClientComponentAdapter(this.registry, this.target, super.element) : super(priority: 1000);
 
-  final String name;
-  final String? data;
+  final ClientComponentRegistryElement registry;
+  final ClientTarget target;
+
+  bool _isClientBoundary = true;
+  late String? data;
+
+  @override
+  void prepareBoundary(ChildListRange range) {
+    element.visitAncestorElements((e) {
+      if (registry.serverElements.containsKey(e)) {
+        return false;
+      } else if (registry.clientElements.containsKey(e)) {
+        _isClientBoundary = false;
+        return false;
+      }
+      return true;
+    });
+
+    if (!_isClientBoundary) {
+      return;
+    }
+
+    data = getData();
+  }
 
   @override
   void applyBoundary(ChildListRange range) {
+    if (!_isClientBoundary) {
+      return;
+    }
+
     final startMarker = MarkupRenderText(
-      '<!--${DomValidator.clientMarkerPrefix}$name${data != null ? ' data=$data' : ''}-->',
+      '<!--${DomValidator.clientMarkerPrefix}${target.name}${data != null ? ' data=$data' : ''}-->',
       true,
     );
-    final endMarker = MarkupRenderText('<!--/${DomValidator.clientMarkerPrefix}$name-->', true);
+    final endMarker = MarkupRenderText('<!--/${DomValidator.clientMarkerPrefix}${target.name}-->', true);
 
     range.start.insertNext(ChildNodeData(startMarker));
     range.end.insertPrev(ChildNodeData(endMarker));
   }
-}
 
-class ClientComponentRegistry extends ObserverComponent {
-  ClientComponentRegistry({required super.child, super.key});
-
-  @override
-  ObserverElement createElement() => ClientComponentRegistryElement(this);
-}
-
-class ClientComponentRegistryElement extends ObserverElement {
-  ClientComponentRegistryElement(super.component);
-
-  bool _didHandleClientScript = false;
-  final List<Element> _clientElements = [];
-  final List<ClientTarget> _clientTargets = [];
-
-  @override
-  void willRebuildElement(Element element) {}
-
-  @override
-  void didRebuildElement(Element element) {
-    final binding = this.binding as ServerAppBinding;
-
-    if (!_didHandleClientScript) {
-      if (binding.options.clientId != null) {
-        (binding).addRenderAdapter(ClientScriptAdapter(binding.options.clientId!));
-      } else if (kDebugMode) {
-        (binding).addRenderAdapter(NoClientScriptAdapter());
-      }
-      _didHandleClientScript = true;
-    }
-
-    final entry = binding.options.clients?[element.component.runtimeType];
-
-    if (entry == null) {
-      return;
-    }
-
-    var isClientBoundary = true;
-    element.visitAncestorElements((e) {
-      return isClientBoundary = !_clientElements.contains(e);
-    });
-
-    if (!isClientBoundary) {
-      return;
-    }
-
-    _clientElements.add(element);
-    _clientTargets.add(entry);
-    binding.addRenderAdapter(ClientComponentAdapter(entry.name, entry.dataFor(element.component), element));
+  String? getData() {
+    final data = target.dataFor(
+      element.component,
+      encode: (o) {
+        if (o is Component) {
+          return getDataForServerComponent(o, element);
+        }
+        return o;
+      },
+    );
+    return data;
   }
 
-  @override
-  void didUnmountElement(Element element) {}
+  String getDataForServerComponent(Component component, Element parent) {
+    final key = (parent, component);
+    if (registry.serverComponents[key] case final s?) {
+      return 's${DomValidator.clientMarkerPrefix}$s';
+    }
+
+    final elements = <Element>[];
+
+    void findElementsFromContext(BuildContext context) {
+      context.visitChildElements((e) {
+        if (identical(e.component, component)) {
+          elements.add(e);
+        }
+        findElementsFromContext(e);
+      });
+    }
+
+    findElementsFromContext(parent);
+
+    if (elements.isEmpty) {
+      print('Warning: Component parameter not used in build method. This will result in empty html on the client.');
+      registry.serverComponents[key] = 0;
+      return 's${DomValidator.clientMarkerPrefix}_';
+    }
+
+    final s = registry.serverComponents[key] = registry.serverElementNum++;
+
+    for (final element in elements) {
+      registry.serverElements[element] = s;
+      binding.addRenderAdapter(ServerComponentAdapter(s, element));
+    }
+
+    return 's${DomValidator.clientMarkerPrefix}$s';
+  }
 }