+++ packages/jaspr/CHANGELOG.md
@@ -1,4 +1,11 @@
-## Unreleased patch
+## Unreleased breaking
+
+- Added **Server Components** to allow for more fine-grained control over the server-side rendered component trees.
+- **Breaking**: Renamed `Component.wrapElement()` to `Component.apply()` and added `ApplyTarget target` parameter used to target specific elements instead of only direct children.
+- **Breaking**: Removed support for `attachBetween` parameter in `ClientAppBinding.attachRootComponent()`, as it is no longer needed.
+- Added stateful server-side reload feature.
+
+## 0.23.3
 
 - Added `detachRootComponent()` to `ComponentsBinding` to cleanly unmount the root component.
 - Updated `testClient` in `jaspr_test` to automatically unmount the root component and clean up `document.body` between test cases.